# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for EchoHub Community Forum",
        "title": "EchoHub Community Forum API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/addNewPost": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Add a new post with title, content, and categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Create post",
                "parameters": [
                    {
                        "description": "Post data",
                        "name": "newPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.PostData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comments": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Retrieve all comments for a post by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Get comments",
                "parameters": [
                    {
                        "description": "Comment fetch data",
                        "name": "commentData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_web_handlers.FetchComments"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/echohub-community_internal_models.Comment"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format or PostID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/confirmLogin": {
            "post": {
                "description": "Authenticate user with username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/confirmRegister": {
            "post": {
                "description": "Register a new user with email, username, and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "newUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.NewUserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
        "/createComment": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Post a new comment to an existing post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Create comment",
                "parameters": [
                    {
                        "description": "New comment data",
                        "name": "newCmntData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_web_handlers.CreateComment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comments created succesfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/deletePost": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Remove an existing post by its ID",
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
        "/like-dislike": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Add or remove a like/dislike from a post or comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interactions"
                ],
                "summary": "Like/Dislike",
                "parameters": [
                    {
                        "description": "Interaction data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Interaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid entity type or JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/maxId": {
            "post": {
                "description": "Get the highest post ID for pagination/client-side tracking",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Get max post ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
        "/newPost": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Render the page for creating a new forum post",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Show new post page",
                "responses": {
                    "200": {
                        "description": "New post page HTML",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
        "/posts": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Retrieve a list of posts based on start ID and category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Get posts",
                "parameters": [
                    {
                        "description": "Post filter",
                        "name": "postsData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/echohub-community_internal_models.Post"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
        "/profileSettings": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Render the profile settings page for authenticated users",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Show profile settings",
                "responses": {
                    "200": {
                        "description": "Profile settings page HTML",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
...
        "/updatePost": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Render the page for editing an existing forum post",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Show update post page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update post page HTML",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
        "/updateProfile": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update user's nickname, email, or other details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update profile",
                "parameters": [
                    {
                        "description": "Updated profile data",
                        "name": "toUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.NewUserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Error"
                        }
                    }
                }
            }
        },
        "/updatingPost": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Apply changes to title, content, or categories of a post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Update post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "ID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Updated post data",
                        "name": "toUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.PostData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echohub-community_internal_models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cmd_web_handlers.CreateComment": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "UserID  string `json:\"userid\"`",
                    "type": "string"
                },
                "postid": {
                    "type": "string"
                }
            }
        },
        "cmd_web_handlers.FetchComments": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                }
            }
        },
        "echohub-community_internal_models.Comment": {
            "type": "object",
            "properties": {
                "Content": {
                    "type": "string"
                },
                "CreationDate": {
                    "type": "string"
                },
                "DislikeCount": {
                    "type": "integer"
                },
                "ID": {
                    "type": "integer"
                },
                "LikeCount": {
                    "type": "integer"
                },
                "PostID": {
                    "type": "integer"
                },
                "ProfileImg": {
                    "type": "string"
                },
                "Reaction": {
                    "type": "string"
                },
                "UserID": {
                    "type": "integer"
                },
                "UserName": {
                    "type": "string"
                }
            }
        },
        "echohub-community_internal_models.Error": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                },
                "SubMessage": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "User": {
                    "$ref": "#/definitions/echohub-community_internal_models.User"
                }
            }
        },
        "echohub-community_internal_models.Interaction": {
            "type": "object",
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "entityType": {
                    "type": "string"
                },
                "liked": {
                    "type": "boolean"
                }
            }
        },
        "echohub-community_internal_models.NewUserInfo": {
            "type": "object",
            "properties": {
                "changes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "rPassword": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "echohub-community_internal_models.Post": {
            "type": "object",
            "properties": {
                "Categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "CommentsCount": {
                    "type": "integer"
                },
                "Content": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "DislikeCount": {
                    "type": "integer"
                },
                "ID": {
                    "type": "integer"
                },
                "LikeCount": {
                    "type": "integer"
                },
                "ProfileImg": {
                    "type": "string"
                },
                "Reaction": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "UserName": {
                    "type": "string"
                }
            }
        },
        "echohub-community_internal_models.PostData": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "selectedCategories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "echohub-community_internal_models.Response": {
            "type": "object",
            "properties": {
                "extra": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "echohub-community_internal_models.User": {
            "type": "object",
            "properties": {
                "CreationDate": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "Gender": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "ProfileImg": {
                    "type": "string"
                },
                "UserName": {
                    "type": "string"
                },
                "UserType": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Session-based authentication using cookies",
            "type": "apiKey",
            "name": "userSession",
            "in": "cookie"
        }
    }
}